Phase 2 (D22): restore the PC item storage screen
Errors 184 -> 157. item_pc.c now compiles clean.

+++ include/graphics.h
@@ -3493,7 +3493,7 @@ extern const u32 gBuyMenuFrame_Tilemap[];
 extern const u32 gBuyMenuFrame_TmHmTilemap[];
 extern const u32 gBuyMenuFrame_Palette[];
 extern const u32 gItemPcTiles[];
-extern const u32 gItemPcBgPals[];
+extern const u16 gItemPcBgPals[];
 extern const u32 gItemPcTilemap[];
 extern const u8 gNamingScreenRWindow_Gfx[];
 extern const u8 gNamingScreenPageButton_Gfx[];
